Changes to be committed: 	modified:   tasks/kurpiakov_a_vert_tape_mat_vec_mul/common/include/common.hpp 	modified:   tasks/kurpiakov_a_vert_tape_mat_vec_mul/mpi/include/ops_mpi.hpp 	modified:   tasks/kurpiakov_a_vert_tape_mat_vec_mul/mpi/src/ops_mpi.cpp 	modified:   tasks/kurpiakov_a_vert_tape_mat_vec_mul/seq/src/ops_seq.cpp 	modified:   tasks/kurpiakov_a_vert_tape_mat_vec_mul/tests/functional/main.cpp 	modified:   tasks/kurpiakov_a_vert_tape_mat_vec_mul/tests/performance/main.cpp

+++ tasks/kurpiakov_a_vert_tape_mat_vec_mul/common/include/common.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cstdint>
 #include <string>
 #include <tuple>
 #include <vector>
@@ -10,7 +9,7 @@
 namespace kurpiakov_a_vert_tape_mat_vec_mul {
 
 using InType = std::tuple<int, std::vector<int>, std::vector<int>>;
-using OutType = std::vector<int64_t>;
+using OutType = std::vector<long long>;
 using TestType = std::tuple<InType, std::string, OutType>;
 using BaseTask = ppc::task::Task<InType, OutType>;
