+++ tasks/kurpiakov_a_vert_tape_mat_vec_mul/tests/functional/main.cpp
@@ -2,7 +2,6 @@
 
 #include <array>
 #include <cstddef>
-#include <cstdint>
 #include <string>
 #include <tuple>
 #include <vector>
@@ -54,28 +53,26 @@ TEST_P(KurpiakovARunFuncTestsProcesses2, MatmulFromPic) {
 }
 const std::array<TestType, 10> kTestParam = {
     {std::make_tuple(std::make_tuple(0, std::vector<int>{}, std::vector<int>{}), "test1_empty",
-                     std::vector<int64_t>{}),
+                     std::vector<long long>{}),
      std::make_tuple(std::make_tuple(1, std::vector<int>{2}, std::vector<int>{3}), "test2_1x1",
-                     std::vector<int64_t>{6}),
-     std::make_tuple(std::make_tuple(2, std::vector<int>{1, 2, 3, 4}, std::vector<int>{5, 6}),
-                     "test3_2x2", std::vector<int64_t>{17, 39}),
-     std::make_tuple(std::make_tuple(3, std::vector<int>{1, 0, 0, 0, 1, 0, 0, 0, 1},
-                                     std::vector<int>{2, 3, 4}),
-                     "test4_identity_3x3", std::vector<int64_t>{2, 3, 4}),
-     std::make_tuple(std::make_tuple(2, std::vector<int>{0, 0, 0, 0}, std::vector<int>{1, 2}),
-                     "test5_zero_matrix", std::vector<int64_t>{0, 0}),
-     std::make_tuple(std::make_tuple(2, std::vector<int>{1, 2, 3, 4}, std::vector<int>{0, 0}),
-                     "test6_zero_vector", std::vector<int64_t>{0, 0}),
+                     std::vector<long long>{6}),
+     std::make_tuple(std::make_tuple(2, std::vector<int>{1, 2, 3, 4}, std::vector<int>{5, 6}), "test3_2x2",
+                     std::vector<long long>{17, 39}),
+     std::make_tuple(std::make_tuple(3, std::vector<int>{1, 0, 0, 0, 1, 0, 0, 0, 1}, std::vector<int>{2, 3, 4}),
+                     "test4_identity_3x3", std::vector<long long>{2, 3, 4}),
+     std::make_tuple(std::make_tuple(2, std::vector<int>{0, 0, 0, 0}, std::vector<int>{1, 2}), "test5_zero_matrix",
+                     std::vector<long long>{0, 0}),
+     std::make_tuple(std::make_tuple(2, std::vector<int>{1, 2, 3, 4}, std::vector<int>{0, 0}), "test6_zero_vector",
+                     std::vector<long long>{0, 0}),
      std::make_tuple(std::make_tuple(2, std::vector<int>{1, -2, -3, 4}, std::vector<int>{2, 1}),
-                     "test7_negative_elements", std::vector<int64_t>{0, -2}),
-     std::make_tuple(std::make_tuple(2, std::vector<int>{1000000, 2000000, 3000000, 4000000},
-                                     std::vector<int>{500, 600}),
-                     "test8_large_values", std::vector<int64_t>{1700000000, 3900000000}),
-     std::make_tuple(std::make_tuple(3, std::vector<int>{2, 1, 1, 1, 2, 1, 1, 1, 2},
-                                     std::vector<int>{1, 1, 1}),
-                     "test9_symmetric_matrix", std::vector<int64_t>{4, 4, 4}),
+                     "test7_negative_elements", std::vector<long long>{0, -2}),
+     std::make_tuple(
+         std::make_tuple(2, std::vector<int>{1000000, 2000000, 3000000, 4000000}, std::vector<int>{500, 600}),
+         "test8_large_values", std::vector<long long>{1700000000LL, 3900000000LL}),
+     std::make_tuple(std::make_tuple(3, std::vector<int>{2, 1, 1, 1, 2, 1, 1, 1, 2}, std::vector<int>{1, 1, 1}),
+                     "test9_symmetric_matrix", std::vector<long long>{4, 4, 4}),
      std::make_tuple(std::make_tuple(2, std::vector<int>{10, 20, 30, 40}, std::vector<int>{2, 4}),
-                     "test10_different_values", std::vector<int64_t>{100, 220})}};
+                     "test10_different_values", std::vector<long long>{100, 220})}};
 
 const auto kTestTasksList = std::tuple_cat(ppc::util::AddFuncTask<KurpiakovAVretTapeMulMPI, InType>(
                                                kTestParam, PPC_SETTINGS_kurpiakov_a_vert_tape_mat_vec_mul),
